refactor(api): move extended flask_restful Api class to seperate file
