+++ app/api.py
@@ -1,0 +1,48 @@
+from flask_restful import Api
+from werkzeug.http import HTTP_STATUS_CODES
+from werkzeug.exceptions import HTTPException
+from requests.exceptions import ConnectionError
+from marshmallow.exceptions import ValidationError as MarshmallowValidationError
+
+
+class ExtendedFlaskRestfulApi(Api):
+
+    def handle_error(self, error):
+        print('Error: ', error)
+
+        if isinstance(error, HTTPException):
+            details = error.message
+            return self._error_response(error=error, status_code=error.code, details=details)
+
+        if isinstance(error, ConnectionError):
+            details = error.message
+            status_code = 502
+            return self._error_response(error=error, status_code=status_code, details=details)
+
+        if isinstance(error, MarshmallowValidationError):
+            details = error.messages
+            status_code = 400
+            return self._error_response(error=error, status_code=status_code, details=details)
+
+        if not getattr(error, 'message', None):
+            details = 'Server has encountered an unexpected error'
+            status_code = 500
+            return self._error_response(error=error, status_code=status_code, details=details)
+
+        # Handle application specific custom exceptions
+        return jsonify(**error.kwargs), error.http_status_code
+
+    def _error_response(self, error, status_code, details):
+        description = getattr(
+            error, 'description', HTTP_STATUS_CODES.get(status_code, '')
+        )
+
+        _dict = {
+            'error': {
+                'code': status_code,
+                'description': description,
+                'details': details,
+            }
+        }
+
+        return _dict, status_code
